Add README file
Per the instruction of TOP's first real assignment, I've created a
README file that will be used in the site. The README itself
consists of a short summary of some of what I've learned up to this
point. Further, the writing of this message and pushing of it was
also a part of this assignment.