Add Home Page and Ham and Cheese Recipe Page
Added a home page to the website as well as the first of what will be 3
recipe pages, the first of which is a Ham and Cheese sammy. The page
and the others moving forward will contain an image, recipe description,
and the actual recipe and utilize lists and links.